requirements.txt
streamlit
streamlit-autorefresh
pandas
numpy
plotly
websocket-client
statsmodels
scikit-learn
